Moved upload into route
